+++ src/menelaus/change_detection/__init__.py
@@ -6,5 +6,6 @@ detection algorithms presented in this package can detect bi-directional shifts,
 either upward or downward changes in a sequence. 
 """
 
-from menelaus.change_detection.page_hinkley import PageHinkley
+from menelaus.change_detection import adwin
 from menelaus.change_detection.cusum import CUSUM
+from menelaus.change_detection.page_hinkley import PageHinkley
